Ajout d'autorisation des users avant delete et update dans le backend, pour que les user ne puissent agir que sur leurs propres posts

+++ backend/controllers/article.js
@@ -9,6 +9,8 @@ var db = require("../services/mysql.config.js");
 //const fs = require("fs");
 //Importation du middleware de vérification des inputs par express-validator 
 const {body} = require('express-validator');
+//Importation du package qui permet de créer et de vérifier les tokens d'authentification 
+const jwt = require("jsonwebtoken");
 
 //Fonction de validation des inputs pour les requêtes post et put
 exports.validate = (method) => {
@@ -50,29 +52,58 @@ exports.createArticle = (req, res, next) => {
   });
 };
 
+ 
 
-//Fontion qui gère la logique métier de la route PUT (modification d'un article posté par son auteur)
+ //Fontion qui gère la logique métier de la route PUT (modification d'un article posté par son auteur)
  exports.modifyArticle = (req, res, next) => {
-    let sql = `UPDATE Articles SET title = ?, description = ?, subject = ?, lien_web  = ?, imageURL = ?, user_id = ?, date_post = ? WHERE id = ?`;
-    let values = [req.body.title, req.body.description, req.body.subject, req.body.lien_web, req.body.imageURL, req.body.user_id, req.body.date_post, req.params.id]
-    db.query(sql, values, function(err, data, fields) {
+     //Recherche dans la BDD de l'article à modifier
+    let sql = "SELECT * FROM Articles WHERE id = ?";
+    db.query(sql, [req.params.id], function(err, data) {
         if (err) {
             return res.status(400).json({err});
         }
-        res.json({status: 201, data, message: "Article modifié avec succès !"})
+        var articleToModify = data[0];
+        //Comparaison de l'id du user courant avec l'id du user ayant posté l'article
+        if (articleToModify.user_id === req.user.userId || req.user.isAdmin === 1) {
+            let sql = `UPDATE Articles SET title = ?, description = ?, subject = ?, lien_web  = ?, user_id = ?, date_post = ? WHERE id = ?`;
+            let values = [req.body.title, req.body.description, req.body.subject, req.body.lien_web, req.user.userId, req.body.date_post, req.params.id]
+            db.query(sql, values, function(err, data, fields) {
+                if (err) {
+                    return res.status(400).json({err});
+                }
+                res.json({status: 201, data, message: "Article modifié avec succès !"})
+            });
+        } else {
+            throw "La modification de l'article n'est possible que par son auteur.";
+        }
     });
-}; 
+ }
 
+    
+   
         
 //Fontion qui gère la logique métier de la route DELETE (suppression d'un article posté)
 exports.deleteArticle = (req, res, next) => {
-  let sql = "DELETE FROM Articles WHERE id = ?";
-  db.query(sql, [req.params.id], function(err, data) {
-    if (err) {
-        return res.status(400).json({err});
-    }
-    res.json({status: 200, data, message: "Article supprimé avec succès !"})
-  });
+    //Recherche dans la BDD de l'article à modifier
+    let sql = "SELECT * FROM Articles WHERE id = ?";
+    db.query(sql, [req.params.id], function(err, data) {
+        if (err) {
+            return res.status(400).json({err});
+        }
+        var articleToModify = data[0];
+        //Comparaison de l'id du user courant avec l'id du user ayant posté l'article
+        if (articleToModify.user_id === req.user.userId || req.user.isAdmin === 1) {
+              let sql = "DELETE FROM Articles WHERE id = ?";
+              db.query(sql, [req.params.id], function(err, data) {
+                if (err) {
+                    return res.status(400).json({err});
+                }
+                res.json({status: 200, data, message: "Article supprimé avec succès !"})
+              });
+        } else {
+            throw "La suppression de l'article n'est possible que par son auteur.";
+        }
+    });
 };
